Day 9 pt 1 - need to fix 203 error.

+++ src/main/java/thebergers/adventofcode2019/day07/Day07Part1.java
@@ -14,7 +14,7 @@ public class Day07Part1 {
 				.setProgram(program)
 				.setPhaseSettingRange(0, 4)
 				.build();
-		Integer maxThrust = tsc.calculateMaxThrust();
+		Long maxThrust = tsc.calculateMaxThrust();
 		System.out.format("Max thrust signal = %d%n", maxThrust);
 	}
 }
